Metoderna flyttas till EditBookings klassen
